+++ ingenialink/virtual/virtual_drive.py
@@ -1,18 +1,19 @@
+import random
 import socket
 import time
 from enum import Enum
 from threading import Thread
-import random
-from typing import Tuple, List, Dict, Union, Optional
+from typing import Dict, List, Optional, Tuple, Union
 
 from ingenialink.constants import ETH_BUF_SIZE, MONITORING_BUFFER_SIZE
-from ingenialink.utils.mcb import MCB
-from ingenialink.utils._utils import convert_bytes_to_dtype, convert_dtype_to_bytes
-from ingenialink.ethernet.servo import EthernetServo
+from ingenialink.enums.register import REG_ACCESS, REG_DTYPE
 from ingenialink.ethernet.dictionary import EthernetDictionary
 from ingenialink.ethernet.register import EthernetRegister
-from ingenialink.enums.register import REG_DTYPE, REG_ACCESS
+from ingenialink.ethernet.servo import EthernetServo
 from ingenialink.utils import constants
+from ingenialink.utils._utils import convert_bytes_to_dtype, convert_dtype_to_bytes
+from ingenialink.utils.constants import IL_MC_CW_EO
+from ingenialink.utils.mcb import MCB
 
 
 class MSG_TYPE(Enum):
@@ -345,6 +346,7 @@ class VirtualDrive(Thread):
         self._update_registers()
         self.__monitoring = VirtualMonitoring(self)
         self.__disturbance = VirtualDisturbance(self)
+        self.__set_motor_ready_to_switch_on()
         self.socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
 
     def run(self) -> None:
@@ -410,7 +412,6 @@ class VirtualDrive(Thread):
         self.set_value_by_id(1, "PROF_POS_VEL_RATIO", 1.0)
         self.set_value_by_id(1, "FBK_BISS_CHAIN", 1)
         self.set_value_by_id(1, "DRV_PS_FREQ_SELECTION", 0)
-        self.set_value_by_id(1, "DRV_STATE_STATUS", constants.IL_MC_PDS_STA_RTSO)
         self.set_value_by_id(1, "DRV_POS_VEL_RATE", 20000)
         self.set_value_by_id(0, "DIST_MAX_SIZE", 8192)
         self.set_value_by_id(0, "MON_MAX_SIZE", 8192)
@@ -523,6 +524,12 @@ class VirtualDrive(Thread):
             self.__disturbance.remove_data()
         if reg_id == "DIST_DATA" and subnode == 0:
             self.__disturbance.append_data(data)
+        if reg_id == "DRV_STATE_CONTROL" and subnode == 1 and (int(value) & IL_MC_CW_EO):
+            self.__set_motor_enable()
+        if reg_id == "DRV_STATE_CONTROL" and subnode == 1 and (value == constants.IL_MC_PDS_CMD_DV):
+            self.__set_motor_disable()
+        if reg_id == "DRV_STATE_CONTROL" and subnode == 1 and (value == constants.IL_MC_PDS_CMD_SD):
+            self.__set_motor_ready_to_switch_on()
 
     def address_to_id(self, subnode: int, address: int) -> str:
         """Converts a register address into its ID.
@@ -609,3 +616,15 @@ class VirtualDrive(Thread):
             if id is None:
                 raise ValueError("Register address or id should be passed")
         return self.__dictionary.registers(subnode)[id]
+
+    def __set_motor_enable(self) -> None:
+        """Set the enabled state."""
+        self.set_value_by_id(1, "DRV_STATE_STATUS", constants.IL_MC_PDS_STA_OE)
+
+    def __set_motor_disable(self) -> None:
+        """Set the disabled state."""
+        self.set_value_by_id(1, "DRV_STATE_STATUS", constants.IL_MC_PDS_STA_SOD)
+
+    def __set_motor_ready_to_switch_on(self) -> None:
+        """Set the ready-to-switch-on state."""
+        self.set_value_by_id(1, "DRV_STATE_STATUS", constants.IL_MC_PDS_STA_RTSO)